Conflict resolution summary:
- .gitignore: accept upstream (website artifacts + *.deb)
- gpu_pci_device_spec.cpp: take upstream generation-registry dispatch;
  remove kGfx942/kGfx950TargetVersion constants (now in registry)
- gpu_generation_registry.cpp: add gfx942 + gfx950 to kGenerations[]
  (topology from gfx942_cdna3_kmd.json / gfx950_mi355x_kmd.json)
- ip_discovery_profile.h: auto-merged; remove = {} defaults on our
  gfx942/gfx950 declarations (upstream pattern: no default params)
- ip_discovery_profile.cpp: auto-merged (gfx942 + gfx950 + gfx1250)
- ip_discovery_test.cpp: keep our gfx942/gfx950 contract tests; add
  upstream AdvertisesNoMediaBlocks; update calls to use registry-backed
  gfx942_spec() / gfx950_spec() helpers (no-arg form removed)

Upstream structural change integrated: 48e5a39c90
  generation-based GPU discovery - GpuGenerationDescriptor registry,
  GpuDiscoveryTopology with full fields, gpu_generation_registry.h/.cpp

Closes #118 (partial - fork side complete; mivirt submodule bump pending)

+++ projects/rccl/src/algorithms/dda/alltoall/alltoall_dda.h
@@ -48,8 +48,8 @@ __launch_bounds__(512)
 #pragma unroll NRANKS
     for (int r = 0; r < NRANKS; ++r) {
       int srcRank = r;
-      int srcIdx = idx + selfRank * idxEnd;
-      int destIdx = idx + r * idxEnd;
+      size_t srcIdx = idx + selfRank * idxEnd;
+      size_t destIdx = idx + r * idxEnd;
       *reinterpret_cast<uint4*>(&recvbuff[destIdx]) = reinterpret_cast<const uint4*>(&ipcbuffs[srcRank][srcIdx])[0];
     }
   }
